Move logr.hh to common directory. NFC
